Add implementation to classes
GameObject can now have a graphics component. The goal is to make the
graphics component fully in charge of rendering, so the rest of the
class doesn't have to worry about it.

On the way I also got familiar with C's circular depency problem. Also
modified CMakeLists to use c11 (when needed).