Fix GameObject vector
GameObject vector now works! The problem was that std::vector *copies*
the objects that are added to it. Thus objects were copied, and then
destroyed. Now GameObject pointers are copied to the vector, solving the
problem.

From StackOverflow:
"What you have to know about vectors in c++ is that they have to use the
copy operator of the class of your objects to be able to enter them into
the vector. If you had memory allocation in these objects that was
automatically desallocated when the destructor was called, that could
explain your problems : your object was copied into the vector then
destroyed.

If you have, in your object class, a pointer that points towards a
buffer allocated, a copy of this object will point towards the same
buffer (if you use the default copy operator). If the destructor
desallocates the buffer, when the copy destructor will be called, the
original buffer will be desallocated, therefore your data won't be
available anymore.

This problem doesn't happen if you use pointers, because you control the
life of your elements via new/destroy, and the vector functions only
copy pointer towards your elements."

-Tuxer, @StackOverflow