Fix Object3D nullptr issue
GraphicsComponent couldn't transform its graphics object, because the
pointer it gave to loadModel() as an argument was overwritten. (Or
something, I'm not too sure). So the graphics object pointer had to be
allocated in GraphicsComponent.

This however gave rise to a new problem. Because the scenegraph manages
the its children, and scenegraph was deleted before GraphicsComponent
when exit()'ng the program, GraphicsComponent tried to delete a nullptr.

This was solved by calling Logic's cleanup() function before
SDL2::exit(), so that all GameObjects were destroyed first.

Also, because GraphicsComponent doesn't own the graphics object
(Object3D), I've turned it into a bare pointer to avoid confusion.
It still allocates and destroyes it though, which is confusing.